Separate out fixture steps into new context file
