+++ behat.yml.dist
@@ -4,11 +4,15 @@ default:
             contexts:
                 - ApplicationContext:
                     kernel: '@kernel'
+                - FixturesContext:
+                    kernel: '@kernel'
             filters:
                 tags: "@fetch"
         api:
             contexts:
                 - ApiContext:
+                    kernel: '@kernel'
+                - FixturesContext:
                     kernel: '@kernel'
             filters:
                 tags: "@fetch && @api"
