+++ features/bootstrap/ApplicationContext.php
@@ -1,7 +1,5 @@
 <?php
 
-use App\Entity\JoindInEvent;
-use App\Entity\JoindInTalk;
 use App\Repository\JoindInCommentRepository;
 use App\Repository\JoindInEventRepository;
 use App\Repository\JoindInTalkRepository;
@@ -10,7 +8,6 @@ use App\Service\JoindInCommentRetrieval;
 use App\Service\JoindInEventRetrieval;
 use App\Service\JoindInTalkRetrieval;
 use Behat\Behat\Context\Context;
-use Behat\Gherkin\Node\TableNode;
 use Doctrine\ORM\EntityManager;
 use Symfony\Component\HttpKernel\KernelInterface;
 use Webmozart\Assert\Assert;
@@ -31,60 +28,6 @@ class ApplicationContext implements Context
     public function __construct(KernelInterface $kernel)
     {
         $this->kernel = $kernel;
-    }
-
-    /**
-     * @Given there are no meetups in the system
-     */
-    public function thereAreNoMeetupsInTheSystem()
-    {
-        foreach ($this->getEventRepository()->findAll() as $joindInEvent) {
-            foreach ($joindInEvent->getTalks() as $joindInTalk) {
-                foreach ($joindInTalk->getComments() as $joindInComment) {
-                    $this->getEntityManager()->remove($joindInComment);
-                }
-
-                $this->getEntityManager()->remove($joindInTalk);
-            }
-
-            $this->getEntityManager()->remove($joindInEvent);
-        }
-        $this->getEntityManager()->flush();
-    }
-
-    /**
-     * @Given we have these meetups in the system
-     */
-    public function weHaveThisMeetupsInTheSystem(TableNode $table)
-    {
-        $this->thereAreNoMeetupsInTheSystem();
-
-        foreach ($table as $row) {
-            $event = new JoindInEvent((int) $row['id'], $row['title'], new DateTime($row['startDate']), new DateTime($row['endDate']));
-
-            $this->getEntityManager()->persist($event);
-        }
-        $this->getEntityManager()->flush();
-    }
-
-    /**
-     * @Given we have these talks in the system
-     */
-    public function weHaveThisTalksInTheSystem(TableNode $table)
-    {
-        foreach ($table as $row) {
-            $event = $this->getEventRepository()->find($row['eventId']);
-
-            $talk = new JoindInTalk(
-                (int) $row['id'],
-                $row['title'],
-                $event,
-                new DateTime($row['importedAt'])
-            );
-
-            $this->getEntityManager()->persist($talk);
-        }
-        $this->getEntityManager()->flush();
     }
 
     /**
